feat: end-to-end CLI pipeline (Epic 7)
- cli.py: wave-from-video <video> runs the full pipeline (overlay, npz,
  waveform plot, spatial.wav, temporal.wav); picks the strongest-wiggle
  frame as representative
- scripts/run_pipeline.sh wrapper
- store 2D arrays as float32 in the npz (7.5MB)
- commit sample outputs as deliverables
- CLI integration test on 12 real frames

Full run: 2642 frames in ~45s. Verified: 21 tests pass, ruff clean.

+++ src/wave_from_video/render.py
@@ -102,8 +102,10 @@ def save_waveform_npz(
         out_path,
         spatial_x=spatial_x,
         spatial_amplitude=spatial_amplitude,
-        centerlines=temporal.centerlines,
-        amplitudes=temporal.amplitudes,
+        # 2-D raw arrays stored as float32 to keep the file small; the 1-D
+        # signals that drive plots/audio stay float64.
+        centerlines=temporal.centerlines.astype(np.float32),
+        amplitudes=temporal.amplitudes.astype(np.float32),
         envelope=temporal.envelope,
         fps=temporal.fps,
         stride=temporal.stride,
